requirements.txt
PyQt6~=6.3.1
Pillow~=9.2.0